qr_menu_backend/requirements.txt
fastapi
uvicorn
uvicorn[standard]
openai
python-dotenv
pyjwt
python-multipart
